requirements.txt
SQLAlchemy~=2.0.29
Flask-Login~=0.6.3
Werkzeug~=3.1.3
Flask~=3.1.0
Flask-RESTful~=0.3.10
pillow~=11.2.1